MiniPi

Documentation for MiniPi.

A miniature program that can compute Pi to millions of digits. —— Mini-Pi in Pure Julia.

function index

MiniPi.EXTRA_PRECISION

MiniPi.WORD_MAX

MiniPi.WORD_SIZE

MiniPi.log2pi_p1d2

MiniPi.MiniBf

MiniPi.MiniBf

Base.Math.exponent

Base.div

Base.iszero

Base.precision

Base.sign

MiniPi.Exp

MiniPi.FFT.ensure_fft_tables

MiniPi.FFT.ensure_fft_tables

MiniPi.FFT.fft_forward!

MiniPi.FFT.fft_inverse!

MiniPi.FFT.fft_pointwise!

MiniPi.FFT.fft_to_word!

MiniPi.FFT.minimum_fft_size

MiniPi.FFT.multiply_fft!

MiniPi.FFT.word_to_fft!

MiniPi.Pi

MiniPi._magnitude

MiniPi.add

MiniPi.e_bsr

MiniPi.e_terms

MiniPi.invsqrt

MiniPi.logf_approx

MiniPi.mul

MiniPi.mul

MiniPi.negate!

MiniPi.pi_bsr

MiniPi.rcp

MiniPi.sub

MiniPi.to_string

MiniPi.to_string_sci!

MiniPi.to_string_trimmed!

MiniPi.uadd

MiniPi.ucmp

MiniPi.usub

MiniPi.word_at

MiniPi.zero!

function docs

MiniPi.EXTRA_PRECISION — Constant

Precision increase step.

MiniPi.WORD_MAX — Constant

Word max

MiniPi.WORD_SIZE — Constant

Word size = 10^9

MiniPi.log2pi_p1d2 — Constant

(log(2*pi) + 1) / 2

MiniPi.MiniBf — Type

MiniBf(x::UInt32, sign=true)

Construct a BigFloat with a value of x and the specified sign.

MiniPi.MiniBf — Type

The Big Floating-point object. It represents an arbitrary precision floating-point number.

Its numerical value is equal to:

    word = 10^9
    word^exp * (T[0] + T[1]*word + T[2]*word^2 + ... + T[L - 1]*word^(L - 1))

T is an array of 32-bit integers. Each integer stores 9 decimal digits and must always have a value in the range [0, 999999999].

T[L - 1] must never be zero.

The number is positive when (sign = true) and negative when (sign = false). Zero is represented as (sign = true) and (len = 0).

Base.Math.exponent — Method

Base.exponent(x::MiniBf)

Returns the exponent of the number in words.

Note that each word is 9 decimal digits.

Base.div — Method

Base.div(x::MiniBf, y::UInt32, p::UInt64)

Division

Depend on: mul, rcp

Base.iszero — Method

If x is zero.

Base.precision — Method

Base.precision(x::MiniBf)

Returns the precision of the number in words.

Note that each word is 9 decimal digits.

Base.sign — Method

Return x sign.

MiniPi.Exp — Function

Compute e using the Taylor Series of exp(1).

MiniPi.Pi — Function

Compute Pi using the Chudnovsky Formula.

MiniPi._magnitude — Method

magnitude(x::MiniBf)

Get magnitude.

MiniPi.add — Function

add(x::MiniBf, y::MiniBf, p=zero(UInt64))

Addition

The target precision is p.

If (p = 0), then no truncation is done. The entire operation is done at maximum precision with no data loss.

MiniPi.e_bsr — Method

Binary Splitting recursion for exp(1).

MiniPi.e_terms — Method

Returns the # of terms needed to reach a precision of p.

The taylor series converges to log(x!) / log(10) decimal digits after x terms. So to find the number of terms needed to reach a precision of p we need to solve this question for x:

p = log(x!) / log(1_000_000_000)

This function solves this equation via binary search.

MiniPi.invsqrt — Method

invsqrt(x::UInt32, p::UInt64)

Compute inverse square root using Newton's Method.

            (  r0^2 * x - 1  )
  r1 = r0 - (----------------) * r0
            (       2        )

Depend on: invsqrt, sub, mul

MiniPi.logf_approx — Method

logf_approx(x::Float64)

Returns a very good approximation to log(x!).

log(x!) ~ (x + 1/2) * (log(x) - 1) + (log(2*pi) + 1) / 2

This approximation gets better as x is larger.

MiniPi.mul — Function

mul(x::MiniBf, y::MiniBf, p=zero(UInt64))

Multiplication

The target precision is p. If (p = 0), then no truncation is done. The entire operation is done at maximum precision with no data loss.

Depend on: multiply_fft!

MiniPi.mul — Method

mul(x::MiniBf, y::UInt32)

Multiply by a 32-bit unsigned integer.

MiniPi.negate! — Method

negate!(x::MiniBf)

Negate this number.

MiniPi.pi_bsr — Method

Binary Splitting recursion for the Chudnovsky Formula.

MiniPi.rcp — Method

rcp(x::MiniBf, p::UInt64)

Compute reciprocal using Newton's Method.

r1 = r0 - (r0 * x - 1) * r0

Depend on: rcp, sub, mul

MiniPi.sub — Function

sub(x::MiniBf, y::MiniBf, p=zero(UInt64))

Subtraction: x - y, The target precision is p.

If (p = 0), then no truncation is done. The entire operation is done at maximum precision with no data loss.

MiniPi.to_string — Function

to_string(x::MiniBf, to_digits::Int64)

Convert this number to a string. Auto-select format type.

MiniPi.to_string_sci! — Function

to_string_sci!(u8::Vector{UInt8}, x::MiniBf, to_digits::Int64)

Convert to string in scientific notation.

MiniPi.to_string_trimmed! — Method

to_string_trimmed!(str::Vector{UInt8}, x::MiniBf, to_digits::Int64)

Converts this object to a string with "digits" significant figures.

After calling this function, the following expression is equal to the numeric value of this object. (after truncation of precision) str + " * 10^" + (return value)

MiniPi.uadd — Function

uadd(x::MiniBf, y::MiniBf, p=zero(UInt64))

Perform addition ignoring the sign of the two operands.

MiniPi.ucmp — Method

ucmp(x::MiniBf, y::MiniBf)

Compare function that ignores the sign.

This is needed to determine which direction subtractions will go.

MiniPi.usub — Function

usub(xx::MiniBf, yy::MiniBf, p=zero(UInt64))

Perform subtraction ignoring the sign of the two operands.

"this" must be greater than or equal to y. Otherwise, the behavior is undefined.

MiniPi.word_at — Method

word_at(x::MiniBf, mag::Int64)

Returns the word at the mag'th digit place.

This is useful for additions where you need to access a specific "digit place" of the operand without having to worry if it's out-of-bounds.

This function is mathematically equal to:

(return value) = floor(this * (10^9)^-mag) % 10^9

MiniPi.zero! — Method

Set x to zero.

MiniPi.FFT.ensure_fft_tables — Function

Ensure that the pre-computed twiddle factor table is large enough to handle a product size of CL.

This method is not thread-safe with "multiply_FFT()".

MiniPi.FFT.ensure_fft_tables — Method

No-Op for basic impl.

MiniPi.FFT.fft_forward! — Method

fft_forward!(T::Vector{ComplexF64}, k::Int)

This function performs a forward FFT of length 2^k.

This is a Decimation-in-Frequency (DIF) FFT. The frequency domain output is in bit-reversed order.

Parameters:

T: Pointer to array.

k: 2^k is the size of the transform

MiniPi.FFT.fft_inverse! — Method

fft_inverse(T::AbstractVector{ComplexF64}, k::Int)

This function performs an inverse FFT of length 2^k.

This is a Decimation-in-Time (DIT) FFT. The frequency domain input must be in bit-reversed order.

Parameters:

T: Pointer to array.

k: 2^k is the size of the transform

MiniPi.FFT.fft_pointwise! — Method

fft_pointwise(T::Vector{ComplexF64}, A::Vector{ComplexF64}, k::Int)

Performs pointwise multiplications of two FFT arrays.

Parameters:

T: Pointer to array.

k: 2^k is the size of the transform

MiniPi.FFT.fft_to_word! — Function

fft_to_word!(A::Vector{UInt32}, AL::UInt64, T::Vector{ComplexF64}, k::Int)

Convert FFT array back to word array. Perform rounding and carryout.

Parameters:

A: word array

AL: length of word array

T: FFT array

k: 2^k is the size of the transform

MiniPi.FFT.minimum_fft_size — Method

Determine minimum FFT size.

MiniPi.FFT.multiply_fft! — Function

multiply_fft!(
    C::Vector{UInt32},
    A::Vector{UInt32}, AL::UInt64,
    B::Vector{UInt32}, BL::UInt64,
)

Multiply A by B and store into C.

AL and BL are the lengths of A and B.

AL + BL is the length of C.

A, B, and C point the start of a little-endian array big integer.

MiniPi.FFT.word_to_fft! — Method

word_to_fft(T::Vector{ComplexF64}, k::Int, A::Vector{UInt32}, AL::UInt64)

Convert word array into FFT array. Put 3 decimal digits per complex point.

Parameters:

T: FFT array

k: 2^k is the size of the transform

A: word array

AL: length of word array